Refactor database migration and seeding process
Updating the database migration and seeding process to favor TS files.
This standard provides a common scripting mechanism, and ensures that
the type safety of TS is leveraged in our scripting files.

In addition to the scripting enhancements, I've also created
generator functions to handle creating random seed data for the
database.

I created a data type interface for Advocate that matches our database schema
to facilitate the passing of data between UI and API layers.

I've installed TSX to allow us to run TypeScript files, which should
streamline scripting.