Learning drizzle and making some changes.
After learning some things about drizzle, I've refactored
the npm scritps to use the drizzle CLI instead of the migrate.ts script.

Preferred using inferred types for the schema files instead of
keeping interfaces in sync with the schema.

I noticed the column for specialties was called payload. I didn't think
this name was fitting or consistent, so I renamed it to specialties to ensure
consistency in the DB and the code.