Implementing search capability for advocates.
In the next config file, I've turned off strict mode to disable
double rendering.

I added a toast package for sending notifications to the user,
particularly in the case that the API sends a non-200 response.

I am intentionally not styling the page in this PR. This PR
strictly handles implementing the search capabilities.