Feature: Relational DB structure
Moving the advocates and specialties to normalized data structure.
I got a little tired of dealing with the casting of jsonb due
to the limitations of drizzle-kit, so I decided to opt for a normalized
data structure.

It comes at a cost of convenience, but I think the normalized data
structure will perform better for querying larger data sets. Also,
the new data structure will support better indexing across the tables,
as that must be a consideration when the data set grows.

Due to the relational data modeling, I opted for interfaces to represent
that data when passing it between the UI and API layers. It's a slight
maintenance overhead, but I think it makes dealing with the data easier
than trying to infer the data types from drizzle.

Despite initially not committing to it in my DISCUSSION.md comments, I
felt it necessary to implement the repository pattern to facilitate the
extraction of records in a clean and reusable way. I defined a base
interface for the repository to follow, which could be expanded upon with
different CRUD methods.