Feature: Unify seeding approaches
Moved seeder to consolidated approach between api and
script.

Implemented the unified function in the API route.

Added delete capabilities to API — most definitely not a real
world practice, but simply here for the sake of demonstration.
Absolutely would not, and must not have seeding or delete capabilities
anywhere near a public facing API.